pcap/requirements.txt
pyshark
kafka-python